--- a/devolutions-gateway/src/api/rdp.rs
+++ b/devolutions-gateway/src/api/rdp.rs
@@ -8,7 +8,8 @@ use tracing::Instrument as _;
 
 use crate::config::Conf;
 use crate::http::HttpError;
-use crate::session::SessionManagerHandle;
+use crate::recording::ActiveRecordings;
+use crate::session::SessionMessageSender;
 use crate::subscriber::SubscriberSender;
 use crate::token::{CurrentJrl, TokenCache};
 use crate::DgwState;
@@ -20,6 +21,7 @@ pub async fn handler(
         jrl,
         sessions,
         subscriber_tx,
+        recordings,
         ..
     }): State<DgwState>,
     ConnectInfo(source_addr): ConnectInfo<SocketAddr>,
@@ -27,8 +29,18 @@ pub async fn handler(
 ) -> Result<Response, HttpError> {
     let conf = conf_handle.get_conf();
 
-    let response =
-        ws.on_upgrade(move |ws| handle_socket(ws, conf, token_cache, jrl, sessions, subscriber_tx, source_addr));
+    let response = ws.on_upgrade(move |ws| {
+        handle_socket(
+            ws,
+            conf,
+            token_cache,
+            jrl,
+            sessions,
+            subscriber_tx,
+            recordings.active_recordings,
+            source_addr,
+        )
+    });
 
     Ok(response)
 }
@@ -38,15 +50,25 @@ async fn handle_socket(
     conf: Arc<Conf>,
     token_cache: Arc<TokenCache>,
     jrl: Arc<CurrentJrl>,
-    sessions: SessionManagerHandle,
+    sessions: SessionMessageSender,
     subscriber_tx: SubscriberSender,
+    active_recordings: Arc<ActiveRecordings>,
     source_addr: SocketAddr,
 ) {
     let stream = crate::ws::websocket_compat(ws);
 
-    let result = crate::rdp_extension::handle(stream, source_addr, conf, &token_cache, &jrl, sessions, subscriber_tx)
-        .instrument(info_span!("rdp", client = %source_addr))
-        .await;
+    let result = crate::rdp_extension::handle(
+        stream,
+        source_addr,
+        conf,
+        &token_cache,
+        &jrl,
+        sessions,
+        subscriber_tx,
+        &active_recordings,
+    )
+    .instrument(info_span!("rdp", client = %source_addr))
+    .await;
 
     if let Err(error) = result {
         error!(client = %source_addr, error = format!("{error:#}"), "RDP failure");