--- a/devolutions-gateway/src/jmux.rs
+++ b/devolutions-gateway/src/jmux.rs
@@ -5,6 +5,7 @@ use crate::subscriber::SubscriberSender;
 use crate::token::JmuxTokenClaims;
 
 use anyhow::Context as _;
+use devolutions_gateway_task::ChildTask;
 use jmux_proxy::JmuxProxy;
 use tap::prelude::*;
 use tokio::io::{AsyncRead, AsyncWrite};
@@ -17,7 +18,6 @@ pub async fn handle(
     sessions: SessionManagerHandle,
     subscriber_tx: SubscriberSender,
 ) -> anyhow::Result<()> {
-    use futures::future::Either;
     use jmux_proxy::{FilteringRule, JmuxConfig};
 
     let (reader, writer) = tokio::io::split(stream);
@@ -58,20 +58,19 @@ pub async fn handle(
     crate::session::add_session_in_progress(&sessions, &subscriber_tx, info, notify_kill.clone()).await?;
 
     let proxy_fut = JmuxProxy::new(reader, writer).with_config(config).run();
-
-    let proxy_handle = tokio::spawn(proxy_fut);
-    tokio::pin!(proxy_handle);
+    let proxy_handle = ChildTask::spawn(proxy_fut);
+    let join_fut = proxy_handle.join();
+    tokio::pin!(join_fut);
 
     let kill_notified = notify_kill.notified();
     tokio::pin!(kill_notified);
 
-    let res = match futures::future::select(proxy_handle, kill_notified).await {
-        Either::Left((Ok(res), _)) => res.context("JMUX proxy error"),
-        Either::Left((Err(e), _)) => anyhow::Error::new(e).context("Failed to wait for proxy task").pipe(Err),
-        Either::Right((_, proxy_handle)) => {
-            proxy_handle.abort();
-            Ok(())
-        }
+    let res = tokio::select! {
+        res = join_fut => match res {
+            Ok(res) => res.context("JMUX proxy error"),
+            Err(e) => anyhow::Error::new(e).context("Failed to wait for proxy task").pipe(Err),
+        },
+        _ = kill_notified => Ok(()),
     };
 
     crate::session::remove_session_in_progress(&sessions, &subscriber_tx, session_id).await?;