# Pre-release hardening

This hardens Devolutions Gateway service before cutting the next release.

## Graceful shutdown

It is now very important to handle service shutdown in a graceful way.

Indeed, ongoing recording manifests should not get lost when restarting the service.

To that end, I figured it was easier to refactor how tasks and futures are handled in Devolutions Gateway.
On current master, listener and task futures are all box-pinned and collected into a vector that is then
wrapped into a `futures::future::SelectAll` that is then polled by the async runtime.

Relevant code:

- https://github.com/Devolutions/devolutions-gateway/blob/3a869f924c1089fc60b3b0ae85a9ce8cf640fd59/devolutions-gateway/src/service.rs#L116
- https://github.com/Devolutions/devolutions-gateway/blob/3a869f924c1089fc60b3b0ae85a9ce8cf640fd59/devolutions-gateway/src/service.rs#L75

This patch is changing this so that:
- a `tokio` task is spawned for each of these futures,
- a new wrapper struct `ChildTask` is used to explicitly handle whether a task should be "detached" or not, and
- all tasks are given a `ShutdownSignal` watcher in order to gracefully terminate themselves.

Since this change may impact performance, I performed a simple benchmark just for sanity.
(We should build a benchmark suite at some point to be more accurate.)

## Basic benchmark

In this benchmark (a stress test, really), [`oha`](https://github.com/hatoo/oha) is used to generate the HTTP load.

### Procedure

Gateway configuration:

```json
{
  "Id": "aa0b2a02-ba9d-4e87-b707-03c6391b86fb",
  "ProvisionerPublicKeyFile": "provisioner.pub.key",
  "ProvisionerPrivateKeyFile": "provisioner.priv.key",
  "Listeners": [
    {
      "InternalUrl": "tcp://*:8080",
      "ExternalUrl": "tcp://*:8080"
    },
    {
      "InternalUrl": "ws://*:7171",
      "ExternalUrl": "wss://*:7171"
    }
  ],
  "LogDirective": "error",
  "__debug__": {
    "disable_token_validation": true
  }
}  
```

Build and run using `cargo run --release`.

The `/jet/heartbeat` route will be used because it’s one of the most used, and is performing
message passing to obtain the number of running sessions (always 0 in this benchmark).

Make sure other applications such a web browsers are closed.

Then, warm up for a few seconds:

```shell
$ oha -z 10s -H "Authorization: Bearer <REDACTED>" http://localhost:7171/jet/heartbeat
```

And run again for 5 minutes with a few additional arguments:

```shell
$ oha -z 5m -c 50 --no-tui -q 15000 --latency-correction --disable-keepalive -H "Authorization: Bearer <REDACTED>" http://localhost:7171/jet/heartbeat
```

`--disable-keepalive` is for disabling HTTP keep-alive (connection reuse), and `--latency-correction`
along with `-q`, the number of query per second desired, to avoid / mitigate the "Coordinated Omission Problem".

### Results against `master`

```
Summary:
  Success rate: 1.0000
  Total:        300.0001 secs
  Slowest:      0.0155 secs
  Fastest:      0.0003 secs
  Average:      0.0015 secs
  Requests/sec: 14999.9157

  Total data:   549.31 MiB
  Size/request: 128 B
  Size/sec:     1.83 MiB

Response time histogram:
  0.000 [1]       |
  0.002 [3410547] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.003 [986302]  |■■■■■■■■■
  0.005 [93311]   |
  0.006 [7806]    |
  0.008 [1157]    |
  0.009 [454]     |
  0.011 [252]     |
  0.012 [83]      |
  0.014 [49]      |
  0.016 [14]      |

Latency distribution:
  10% in 0.0009 secs
  25% in 0.0010 secs
  50% in 0.0014 secs
  75% in 0.0018 secs
  90% in 0.0024 secs
  95% in 0.0028 secs
  99% in 0.0038 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0002 secs, 0.0000 secs, 0.0103 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0036 secs

Status code distribution:
  [200] 4499976 responses
```

### Results against this patch

```
Summary:
  Success rate: 1.0000
  Total:        300.0004 secs
  Slowest:      0.0170 secs
  Fastest:      0.0003 secs
  Average:      0.0013 secs
  Requests/sec: 14999.9229

  Total data:   549.31 MiB
  Size/request: 128 B
  Size/sec:     1.83 MiB

Response time histogram:
  0.000 [1]       |
  0.002 [4079329] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.004 [401430]  |■■■
  0.005 [16952]   |
  0.007 [1142]    |
  0.009 [398]     |
  0.010 [399]     |
  0.012 [205]     |
  0.014 [110]     |
  0.015 [15]      |
  0.017 [2]       |

Latency distribution:
  10% in 0.0009 secs
  25% in 0.0010 secs
  50% in 0.0012 secs
  75% in 0.0016 secs
  90% in 0.0020 secs
  95% in 0.0023 secs
  99% in 0.0032 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0001 secs, 0.0000 secs, 0.0103 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0034 secs

Status code distribution:
  [200] 4499983 responses
```

### Discussion

- Average response time: 0.0015s → 0.0013s
- Slowest request time: 0.0155s → 0.0170s

It seems the performance is not significantly impacted by this change.

Again, the benchmark is really simple and is only performing a simple API request.

That being said, I think this patch can be merged with a fair degree of confidence that no outstanding
regression has been introduced.

## Responsible recording manifest flushing

A single task ("actor") is now responsible for managing manifest files so that we ensure there is no
data race in a sane way.

When shutting down the service, all pending manifest files are flushed (written).